Allows the filebroweser to be used by plugins.
to call you must call using the follwoing format: 

filebrowser.php?CKEditorFuncNum=1&returnid=post-image2&type=images

CKEditorFuncNum      - Normal CKEditor function uses 2, set it to 1 for user functions
returnid             - Set to the ID of the input field you wish to return the selected image to.
type                 - set to images. 

Sample usage : 

<input onClick='window.open("../admin/filebrowser.php?CKEditorFuncNum=1&returnid=post-image2&type=images","Filebrowser","width=600,height=500")'  type="text" id="post-image2" name="post-image2" value="" />

When an image is selected the URL is returned to provided input field. 




git-svn-id: https://get-simple-cms.googlecode.com/svn/trunk@447 c6e34258-837a-11de-9f42-234e1f44043b