where I moved the $SITEURL to common didnt work, moved it back.
Consolidated all backend files to not use repeated code.
The major change is that the H1 no longer has the page name in it. This was always repeated text anyway (it was always in the H3 orange text)

The install & setup screens also lost the header and sidebar that were never used on those pages anyway.

git-svn-id: https://get-simple-cms.googlecode.com/svn/trunk@594 c6e34258-837a-11de-9f42-234e1f44043b