Update plugin_functions.php
I have created a multi-part plugin, which preserves the boldness of its settings globally. Unfortunately, it is sometimes the case that the base plugin which provides constants etc. simply is not loaded. The sort command of PHP worerst corrected the problem. Beasser Währe a solution via the PHP file and a folder contained therein statement. It could be global files with settings and paths invite ignoring filenames. One possibility would be to extend the function register_plugin.