filter arguments, secfilter variable arguments
you can now pass user args to filter callbacks, not sure if this ever
worked properly or not

i also added variable arguments to secfilters, so we can pass variable
args to callbacks , and users can also, numargs param provided so you do
not haev to use arrays of args, which can get confusing. slightly
experimental, could be expanded to actions as well.