more update item support, testing sorting menu cache during output loop
It might be easier and faster to sort menus output per heirarchy than to
resort the entire pagesarray and rebuild the tree.

Sorting throws a question into whether I chose the best method of doing
this, but tests with over 500 pages shows nested recursion is essential
for speed , and regenerating trees is not efficient. Will keep testing