bunch of crap
checking in for posterity before i remove this entire spaghetti load of
crap i wrote.
Trying to manipulate a nested array in real time using reference maps is
a pain in the ass, i am just going to regenerate it everytime, or cache
it when its not changing, I am removing all this, and maybe rewriting
later using flat mapping array maybe, or using a parent hash bucket and
rebuilding from that so its slightly faster. screw this.